+++ packages/js/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @postrun/js
 
+## 2.7.0
+
 ## 2.6.0
 
 ### Minor Changes